% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny web application}
\usage{
launch_app()
}
\value{
run shiny web application
}
\description{
Launch shiny web application
}
\examples{
\dontrun{launch_app()}


}
