% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_element.R
\name{server_element}
\alias{server_element}
\title{Server Element (load data) of the shiny web application}
\usage{
server_element(file)
}
\arguments{
\item{file}{A csv file to be loaded}
}
\value{
One part of the server of the shiny web application: load data
}
\description{
Server Element (load data) of the shiny web application
}
\examples{
server_element("var1,var2\n  'a',1\n  'b',2")

}
