% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_element.R
\name{ui_element}
\alias{ui_element}
\title{UI Element (a siderbar panel) of the shiny web application}
\usage{
ui_element()
}
\value{
One part of the UI of the shiny web application: a sidebar panel that can be used for month selection
}
\description{
UI Element (a siderbar panel) of the shiny web application
}
\examples{
ui_element()


}
